+++ src/components/cart-dropdown/cart-dropdown.component.jsx
@@ -1,4 +1,6 @@
-import React from "react";
+import React, { useContext } from "react";
+import { CartContext } from "../../context/cart.context";
+import CartItem from "../cart-item/cart-item.component";
 
 import {
   CartDropdownContainer,
@@ -7,10 +9,16 @@ import {
 } from "./cart-dropdown.styles";
 
 const CartDropdown = () => {
+  const { cart } = useContext(CartContext);
+
   return (
     <CartDropdownContainer>
       <CartItems>
-        <EmptyMessage>Your cart is empty</EmptyMessage>
+        {cart.length ? (
+          cart.map((item) => <CartItem key={item.id} cartItem={item} />)
+        ) : (
+          <EmptyMessage>Your cart is empty</EmptyMessage>
+        )}
       </CartItems>
     </CartDropdownContainer>
   );